#106 refactor: TextArea 하위 컴포넌트 분리
- Comment 수정 시 하위 컴포넌트만 재사용하기 위해 분리
